restarting machine

+++ complete/src/main/java/com/example/messagingstompwebsocket/Database/PersonRepository.java
@@ -2,7 +2,7 @@ package com.example.messagingstompwebsocket.Database;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PersonRepository extends JpaRepository<Person, Integer> {
-
+public interface PersonRepository extends JpaRepository<Person, Long> {
 
 }
+
